Index all routes in the file router/index.js and import on server.js
